test: community admin can edit community
