+++ tests-functional/clients/wallet_pending_tx_status.py
@@ -1,0 +1,9 @@
+"""Wallet pending transaction status — mirrors services/wallet/activity/common/types.go TxStatus."""
+
+from enum import Enum
+
+
+class WalletPendingTxStatus(str, Enum):
+    PENDING = "Pending"
+    SUCCESS = "Success"
+    FAILED = "Failed"
